requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
Django==6.0
django-admin-honeypot==1.1.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-ninja==1.5.1
django-ratelimit==4.1.0
django-ratelimit2==0.4.2
django-unfold==0.74.1
pillow==12.0.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-decouple==3.8
qrcode==8.2
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
